Refactor STACK implementations to not expose the IP module or type.
Don't expose the IP module given to stacks in modules fulfilling
V1_LWT.STACKV4.  This allows the socket stacks not to provide
implementations for V1_LWT.IP which will just fail at runtime.  As part
of this change, it's now conceivable to have an implementation of STACK
which incorporates IPV6 rather than IPV4, so rename tcpip_stack_direct
to tcpip_stackv4_direct to avoid possible confusion on the
implementation reflected here.