returns Error `Closed when writing to closed flow
Previously, an attempt to write to a flow which had been closed
would get an Error `Msg "the flow is not yet ready", which was both
too general a polyvar and very misleading.  Return Error `Closed
when the state of the pcb indicates that a connection was once active,
but has been closed.

One would hope that the remaining states are unreachable in writefn, as
the user should only receive a flow to which to write once the
connection has been established.